+++ lib/src/widgets/widgets.dart
@@ -1,9 +1,11 @@
+export 'action_row.dart';
 export 'clamp.dart';
+export 'flap.dart';
 export 'header_bar_minimal.dart';
 export 'header_bar.dart';
 export 'header_button.dart';
 export 'popup_menu.dart';
 export 'sidebar.dart';
-export 'flap.dart';
+export 'preferences_group.dart';
 export 'view_switcher.dart';
 export 'view_switcher_tab.dart';