+++ CHANGELOG.md
@@ -3,5 +3,6 @@
 - Seperate libadwaita (Widgets) and adwaita (Colors)
 - Rename Every widget from GtkSomething to AdwSomething
 - Rename GtkContainer to AdwClamp and GtkTwoPane to AdwFlap
+- Add `AdwPreferenceGroup` and `AdwActionRow` from libadwaita.
 
 For older Changelog visit: https://pub.dev/packages/gtk/changelog